#1597 fix activitity feeds for pull requests
