+++ conf/locale/locale_en-US.ini
@@ -1064,6 +1064,8 @@ create_issue = `opened issue <a href="%s/issues/%s">%s#%[2]s</a>`
 close_issue = `closed issue <a href="%s/issues/%s">%s#%[2]s</a>`
 reopen_issue = `reopened issue <a href="%s/issues/%s">%s#%[2]s</a>`
 create_pull_request = `created pull request <a href="%s/pulls/%s">%s#%[2]s</a>`
+close_pull_request = `closed pull request <a href="%s/pulls/%s">%s#%[2]s</a>`
+reopen_pull_request = `reopened pull request <a href="%s/pulls/%s">%s#%[2]s</a>`
 comment_issue = `commented on issue <a href="%s/issues/%s">%s#%[2]s</a>`
 merge_pull_request = `merged pull request <a href="%s/pulls/%s">%s#%[2]s</a>`
 transfer_repo = transfered repository <code>%s</code> to <a href="%s">%s</a>
