+++ gogs.go
@@ -17,7 +17,7 @@ import (
 	"github.com/gogits/gogs/modules/setting"
 )
 
-const APP_VER = "0.8.57.0305"
+const APP_VER = "0.8.58.0305"
 
 func init() {
 	runtime.GOMAXPROCS(runtime.NumCPU())
