+++ templates/.VERSION
@@ -1,1 +1,1 @@
-0.8.57.0305
+0.8.58.0305
